Moved the API logic for the home page components from the individual presentational components into the Home Container where all logic must reside, and kept all the styling within the individual components. Need to figure out how to filter the data I am receiving from the API by the media type
